WIP: assertlint type casting and redundantcast analyzer
- Add castableType() to only cast basic types (no imports needed)
- Add RedundantCastAnalyzer to detect/fix unnecessary casts like int(0)
- Fix OSC 8 links to use vscode:// protocol with line:column
- Fix several test files with auto-fixer

Known issues:
- Unused import detection incorrectly identifies imports when package
  name differs from path (e.g., go-set exports 'set' not 'go-set')
- Offset-based file modifications are fragile

+++ src/bench/report_test.go
@@ -29,10 +29,10 @@ func TestParseBenchmarkOutput(t *testing.T) {
 
 	foo := results[0]
 	assert.Equal(t, "BenchmarkFoo-8", foo.Name)
-	assert.Equal(t, 10000, foo.Iterations)
-	assert.Equal(t, 123456, foo.NsPerOp)
-	assert.Equal(t, 1024, foo.BytesPerOp)
-	assert.Equal(t, 8, foo.AllocsPerOp)
+	assert.Equal(t, int64(10000), foo.Iterations)
+	assert.Equal(t, float64(123456), foo.NsPerOp)
+	assert.Equal(t, int64(1024), foo.BytesPerOp)
+	assert.Equal(t, int64(8), foo.AllocsPerOp)
 }
 
 func TestParseBenchmarkOutputNoAllocStats(t *testing.T) {
@@ -44,8 +44,8 @@ func TestParseBenchmarkOutputNoAllocStats(t *testing.T) {
 	results := report.Packages["pkg"]
 	require.Equal(t, 1, len(results))
 
-	assert.Equal(t, 0, results[0].BytesPerOp)
-	assert.Equal(t, 0, results[0].AllocsPerOp)
+	assert.Equal(t, int64(0), results[0].BytesPerOp)
+	assert.Equal(t, int64(0), results[0].AllocsPerOp)
 }
 
 func TestParseBenchmarkOutputEmpty(t *testing.T) {
